midi parser fixes

+++ src/core/parseutils.hpp
@@ -11,7 +11,7 @@ namespace MgCore
         T output;
         auto size = sizeof(T);
         int offset;
-        for (int i = 0; i < size; i++) {
+        for (size_t i = 0; i < size; i++) {
             offset = (size-1) - i; // endianness lol ?
             file.read(reinterpret_cast<char*>(&output)+offset, 1);
         }
@@ -35,13 +35,14 @@ namespace MgCore
         return output;
     }
 
+    // Main purpose is for reading string-like data from the file.
     template<typename T>
     void read_type(std::ifstream &file, T *output, unsigned long length)
     {
         auto size = sizeof(T);
         int offset;
-        for (int j = 0; j < length; j++) {
-            for (int i = 0; i < size; i++) {
+        for (size_t j = 0; j < length; j++) {
+            for (size_t i = 0; i < size; i++) {
                 offset = ((size-1) - i) + (size * j); // endianness lol ?
                 file.read(reinterpret_cast<char*>(output) + offset, 1);
             }
@@ -55,7 +56,7 @@ namespace MgCore
         auto size = sizeof(T);
         int offset;
         uint8_t *val;
-        for (int i = 0; i < size; i++) {
+        for (size_t i = 0; i < size; i++) {
             offset = (size-1) - i; // endianness lol ?
             val = reinterpret_cast<uint8_t*>(&output)+offset;
             *val = file.peek();
@@ -82,14 +83,15 @@ namespace MgCore
         return output;
     }
 
+    // Main purpose is for reading string-like data from the file.
     template<typename T>
-    void peek_type(std::ifstream &file, T *output, unsigned long length)
+    void peek_type(std::ifstream &file, T *output, unsigned int length)
     {
         auto size = sizeof(T);
-        int offset;
+        unsigned int offset;
         uint8_t *val;
-        for (int j = 0; j < length; j++) {
-            for (int i = 0; i < size; i++) {
+        for (size_t j = 0; j < length; j++) {
+            for (size_t i = 0; i < size; i++) {
                 offset = ((size-1) - i) + (size * j); // endianness lol ?
                 val = reinterpret_cast<uint8_t*>(&output)+offset;
                 *val = file.peek();
